Add files via upload
The current research analyzes three models of recurrent neural networks (SimpleRNN, LSTM
and GRU) for the purpose of time series analysis and forecasting of air quality in India for the
years 2015 to 2020. Out of the base models, the Gated Recurrent Unit (GRU) was able to
perform best amongst all the other models. It was also found that the proposed model could
effectively handle the changing dynamics over time which are crucial for environmental
monitoring tasks. Nevertheless, once due adjustments of hyperparameters were objectively set,
the LSTM model became the best alternative. This proves that one’s efforts can go a long way
in enhancing the accuracy and precision of a given model. The study underlines the issue of
model choice and its parameters in the case of environmental data analysis, but these concerns
are of course important for public health and policy as well as the practice due to their effect
on the success of interventions.